+++ juniper/CHANGELOG.md
@@ -18,6 +18,9 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
     - Switched `LocalDateTime` scalars to `yyyy-MM-ddTHH:mm:ss` format in types:
         - `chrono::NaiveDateTime`.
         - `time::PrimitiveDateTime`.
+    - Switched from `Date` scalar to `LocalDate` scalar in types:
+        - `chrono::NaiveDate`.
+        - `time::Date`.
     - Renamed `Url` scalar to `URL` in types:
         - `url::Url`.
     - Renamed `Uuid` scalar to `UUID` in types: